requirements.txt
python-docx==1.2.0
fastapi==0.116.1
langchain_core==0.3.76
langchain_huggingface==0.3.1
langchain_postgres==0.0.15
openai==1.107.2
pdfplumber==0.11.7
pydantic==2.11.9
python-dotenv==1.1.1
sentence-transformers==2.2.2
python-multipart==0.0.20
